[ADD] RotBlock_Script
[ADD] - BugFix

+++ Assets/Scripts/Stage1/RotBlockCtrl.cs
@@ -29,7 +29,7 @@ public class RotBlockCtrl : MonoBehaviour
 
     void RotStart()
     {
-        for(int i = 0; i < TriggerBlocks.Length; i++)
+        for(float i = 0; i < TriggerBlocks.Length; i++)
         {
             Invoke("TriggerOn", i/2);
         }
